pyproject.toml
[project]
name = "gen-ai-rag-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0.0"
dependencies = [
    "bs4 (>=0.0.2,<0.0.3)",
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "langchain (>=1.2.7,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-google-genai (>=4.2.0,<5.0.0)",
    "langchain-openai (>=1.1.7,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
